Add Rust and C# equivalents of the "Record List" benchmark
