+++ rust-bench/Cargo.toml
@@ -3,10 +3,6 @@ name = "rust-bench"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-name = "rust-bench"
-path = "src/main.rs"
-
 [dependencies]
 tokio = { version = "1", features = ["full"] }
 tokio-postgres = { version = "0.7", features = ["with-chrono-0_4"] }
